Added statsContext and courseRecommendation
